feat(asset-reminder): move asset-reminder in separeate module
