+++ app/modules/asset/fields.ts
@@ -220,20 +220,3 @@ export const advancedAssetIndexFields = () => {
 
   return fields;
 };
-
-export const ASSET_REMINDER_INCLUDE_FIELDS = {
-  teamMembers: {
-    select: {
-      id: true,
-      name: true,
-      user: {
-        select: {
-          firstName: true,
-          lastName: true,
-          profilePicture: true,
-          id: true,
-        },
-      },
-    },
-  },
-} satisfies Prisma.AssetReminderInclude;
